refactor(defi): share tab models and fiat conversion across Thorchain and Maya
The Thorchain and Maya position ViewModels each carried their own copy of
createFiatValue, identical down to the CancellationException handling, so a
pricing fix had to be made twice. The Bonded/Staking/LP tab models also lived
inside ThorchainDefiPositionsViewModel.kt, which made them look chain-specific
even though the Maya screen renders the same tabs from the same types.

Both now share DefiTabUiModels.kt and an injected DefiFiatValueCalculator. The
models move within their package, so no import changes anywhere.

Also replaces the hardcoded Dispatchers.IO in both ViewModels with the
@IoDispatcher qualifier already used by ten other ViewModels. Same dispatcher in
production; it makes the load paths deterministic under test, which the
accompanying characterization suites depend on.

Deliberately left alone: calculateStakingFiatPrice looks like a twin but
Thorchain catches only java.io.IOException while Maya catches Exception, and
calculateBondedFiatPrice differs by coin and decimal scale. Unifying either
would change Thorchain's failure behavior, so it belongs in its own change.
